// src/types/MonsterType.ts
export interface MonsterType {
    name: string;
    size: string;
    alignment: string;
    type: string;
    cr: string;
    ac: string;
    hp: string;
    speed: string;
    str: string;
    dex: string;
    con: string;
    int: string;
    wis: string;
    cha: string;
    PassivePerception: string;
    Skills?: string;
    Languages?: string;
    abilities: string[];
    Actions?: string[];
    LegendaryActions?: string[];
}